Fix setting HW scalar when not intended (#176)
* Fix Walking Wake, Iron Leaves, MightiestMark and Fly Pikachu, fix Tauros-Paldea test sets, add more sets.

* Replace SV encounter PIDIV check with `ITeraRaid9`.

* Remember to fix RoC's Tauros sets.

+++ PKHeX.Core.AutoMod/AutoMod/Legalization/ShowdownEdits.cs
@@ -46,7 +46,7 @@ namespace PKHeX.Core.AutoMod
             if (pk.Nature == set.Nature || set.Nature == -1)
                 return;
             var val = Math.Min((int)Nature.Quirky, Math.Max((int)Nature.Hardy, set.Nature));
-            if (pk.Species == (int)Species.Toxtricity)
+            if (pk.Species == (ushort)Species.Toxtricity)
             {
                 if (pk.Form == EvolutionMethod.GetAmpLowKeyResult(val))
                     pk.Nature = val; // StatNature already set
